requirements.txt
aws-cdk-lib==2.143.0
constructs>=10.0.0,<11.0.0
telepot
python-dotenv
git+https://github.com/aws-samples/aws-cdk-lambda-layer-builder.git